Conflicts:
	app/admin/forms.rb
	app/admin/sessions.rb
	app/controllers/forms_controller.rb

+++ app/models/case.rb
@@ -9,9 +9,14 @@ class Case < ActiveRecord::Base
   attr_accessible :session_id, :patient_id
   attr_accessible :session, :patient
 
-  validates_uniqueness_of :position, :scope => :session_id
+  validates_uniqueness_of :position, :scope => :session_id  
 
-  before_destroy do |c|
+  before_destroy do
+    unless form_answer.nil?
+      errors.add :base, 'You cannot delete a case that was answered' 
+      return false
+    end
+
     CaseData.destroy_all(:case_id => c.id)
   end
